Add GaugeFloat64 API

+++ start_test.go
@@ -52,7 +52,7 @@ func Test_GlobalMetrics(t *testing.T) {
 			if got, want := s.keys[0], tt.key; !reflect.DeepEqual(got, want) {
 				t.Fatalf("got key %s want %s", got, want)
 			}
-			if got, want := s.vals[0], tt.val; !reflect.DeepEqual(got, want) {
+			if got, want := s.vals[0], float64(tt.val); !reflect.DeepEqual(got, want) {
 				t.Fatalf("got val %v want %v", got, want)
 			}
 		})
@@ -81,7 +81,7 @@ func Test_GlobalMetrics_Labels(t *testing.T) {
 			if got, want := s.keys[0], tt.key; !reflect.DeepEqual(got, want) {
 				t.Fatalf("got key %s want %s", got, want)
 			}
-			if got, want := s.vals[0], tt.val; !reflect.DeepEqual(got, want) {
+			if got, want := s.vals[0], float64(tt.val); !reflect.DeepEqual(got, want) {
 				t.Fatalf("got val %v want %v", got, want)
 			}
 			if got, want := s.labels[0], tt.labels; !reflect.DeepEqual(got, want) {
@@ -123,7 +123,7 @@ func Test_GlobalMetrics_DefaultLabels(t *testing.T) {
 			if got, want := s.keys[0], tt.key; !reflect.DeepEqual(got, want) {
 				t.Fatalf("got key %s want %s", got, want)
 			}
-			if got, want := s.vals[0], tt.val; !reflect.DeepEqual(got, want) {
+			if got, want := s.vals[0], float64(tt.val); !reflect.DeepEqual(got, want) {
 				t.Fatalf("got val %v want %v", got, want)
 			}
 			if got, want := s.labels[0], tt.labels; !reflect.DeepEqual(got, want) {
